libopkg: fix malloc buffer size for replaced string
In the case that users direct opkg to download content from a URL which
uses `ftps://` as its transport scheme, the opkg_download cURL backend
will attempt to replace the scheme with `ftp://`, and allow cURL to
negotiate security parameters.

When the replce_token_in_str() function allocates a new char array for
the replaced string, it erroneously leaves no space for a
null-terminating byte.

Increase the replaced_str_len by 1 when allocating the array, to fix
that bug.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>