libopkg: use memmove for substring operations
GCC 8 throws stringop-truncate warnings when strncpy() operations are
performed on substrings - the reason being that the null-terminating
byte is often lost. The replace_token_in_str() function causes several
warnings for this reason.

Instead, use memmove(), which is functionally similar, but does not
imply that the data it is operating on is a well-formed string.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>