opkg_solver_libsolv: blacklist excluded packages instead of ignoring them
Currently, packages specified with --add-exclude aren't registered at all
with libsolv. Because as far as libsolv is concerned they don't exist, if
an installation requires an excluded package, you get a less-than-ideal
message that "nothing provides" the excluded package.

Instead of pretending that they don't exist, instead utilize libsolv's
recent support for "blacklisted" packages. This changes the error output
to indicate that the excluded package "can only be installed by a direct
request"

Signed-off-by: Brandon Streiff <brandon.streiff@ni.com>
Signed-off-by: Alex Stewart <alex.stewart@ni.com>