opkg_download_curl: relocate return code declaration
When libopkg is compiled with `--enable-curl` AND `--disable-ssl-curl`
options, the `int r` declaration in opkg_curl_init() becomes orphaned
from the code segments which actually use it (which are dropped via
DEFINEs). GCC throws an unused variable warning, as a result.

In modern C, there is no reason that variable declarations need to
happen at the top of the function block. So move the declaration into
the code segment that actually uses it, and within the relevant #DEFINE
block.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>