libopkg: add pre-depends print to package status dumps
opkg installs a package's pre-dependencies but does not record them
in the package's status file. Due to this, opkg allows uninstallation
of an installed package's pre-dependency when it should complain.
This patch fixes that by adding the pre-depends print to the
package's status file. It also adds the pre-depends print to the
package's status info dump.

Given installed packages 'a' and 'b', where 'a'--'pre-depends'-->'b',
the following is the opkg output (after the patch is applied) when
asked to uninstall package 'b' before uninstalling 'a':

DEBUG cmd: /mnt/workspace/src/opkg -o /tmp/opkg  --force-postinstall remove b
DEBUG stdout:  * Solver encountered 1 problem(s):
 * Problem 1/1:
 *   - package a-1.0.all requires b, but none of the providers can be installed
 *   - conflicting requests
 *   - problem with installed package a-1.0.all
 *
 * Solution 1:
 *   - allow deinstallation of a-1.0.all

 * Solution 2:
 *   - do not ask to deinstall b

 *   - do not ask to deinstall b

Signed-off-by: Shruthi Ravichandran <shruthi.ravichandran@ni.com>
Signed-off-by: Alex Stewart <alex.stewart@ni.com>