opkg_conf: check if tmp_dir exists before deleting
rm_r prints an error message if the directory to be deleted does not
exist. There are some opkg commands that do not error, but print the
rm_r error message re: the tmp_dir during cleanup unnecessarily.
Therefore, check if the tmp_dir exists before trying to delete it.

Signed-off-by: Shruthi Ravichandran <shruthi.ravichandran@ni.com>
Signed-off-by: Alex Stewart <alex.stewart@ni.com>