libopkg: remove trailing '/' when generating list from package
In pkg_write_filelist(), the trailing '/' is removed from each entry
before writing to the .list file. But the same is not done when the
.list file doesn't exist yet for a package, and it has to be
generated from the package in pkg_get_installed(). This leads to a
different entry in the file hash for the same file, and multiple
packages claiming ownership of a file.

Fixes commit 70df4de62313cf973e98b296d467be1d73b9b9f9

Signed-off-by: Shruthi Ravichandran <shruthi.ravichandran@ni.com>
Signed-off-by: Alex Stewart <alex.stewart@ni.com>