libopkg: support multiple packages installing a directory or a symlink to one
Multiple packages can install the same directory, or a symlink to a
directory. Ensure that these entries are not removed from the
filelists of packages that are installed following a package that
first installs the directory or a symlink to a directory.
This is later used to ensure that out-of-order package removals do
not result in directory or symlink deletion.

Fixes Bugzilla #10461.

Signed-off-by: Shruthi Ravichandran <shruthi.ravichandran@ni.com>
Signed-off-by: Alex Stewart <alex.stewart@ni.com>