opkg-key: Remove --no-options flag from gpg calls.
The opkg-key script was always passing the --no-options
flag to gpg, which uses /dev/null as the options file.
As a result, the opkg gpg.conf file was not getting
used. This change removes that flag so that gpg.conf
in the GPGHOMEDIR for opkg (currently /etc/opkg/gpg/)
will be used if present.

Signed-off-by: Charlie Johnston <charlie.johnston@ni.com>
Signed-off-by: Alex Stewart <alex.stewart@ni.com>