VersionPoilcy: move to MAINTAINING.md
The VersionPolicy doc describes an approximation of the semantic
versioning policy that this fork follows.

The policy is functionally similar to the popular SemVer spec outlined
here (https://semver.org/spec/v2.0.0.html). So we should formally adopt
that spec to align with community best-practices.

There isn't a compelling reason that the version policy needs to be in
its own file. Create a new MAINTAINING.md document for project
maintenance information and move it there.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>