configure.ac: declare the project as "foreign"
GNU Automake asserts that projects should follow GNU "conventions" and
include some standard files named AUTHORS, INSTALL, ChangeLog, et c. The
intention here is probably to encourage users to distribute this
information, but it is rather heavy-handed to require that they be
specifically named in that fashion.

As a result, automake will error if we try to rename or move these
files.

Declare `-Werror foreign` in the autoconf file, which declares that we
are not a strictly GNU-compliant project in this regard, and squashes
automake's objections.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>