TODO: remove
The TODO file contains some historic musings about things to improve in
the codebase. For most entries, the discussion history has been lost and
the entries themselves aren't enough to take discrete action.

The exception is the request to improve the efficiency of the
pkg_hash_best_installation_candidate function, which has been moved to
this bugzilla.

https://bugzilla.yoctoproject.org/show_bug.cgi?id=15006

Remove the file.

There is some space in the project to keep a roadmap document with
future initiatives that are more than individual enhancements or bugs,
and that document is in-work.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>