Makefile.am: do not duplicate EXTRA_DIST dirs
Automake is not very intelligent. If you add add a directory to the
EXTRA_DIST variable with a trailing slash, then automake will copy the
directory into a the dist archive under a path of itself. So the dist
archive might contain: `/scripts/scripts/...` or `/tests/tests/...`.

Remove the trailing slash to fix this behavior.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>