migrate mailing list to lists.yoctoproject.org
Google Code/Groups is a legacy product which is not being actively
maintained by Google, and it is becoming progressively more difficult to
maintain the opkg-devel list there.

Instead, move the canonical opkg mailing list for the yocto fork to the
lists.yoctoproject.org site, where it can live with the other
contemporary yocto projects.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>