tests: add test for CVE-2020-7982
The opkg-lede fork was affected by a bug where it would not honor
the package index MD5Sum field. As a result, malicious attackers that
have MITM vector access to an opkg system could inject code into
packages, as long as the overall package size does not change.

Add a test to validate that our opkg fork is not vulnerable to this
exploit.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>