libopkg: Simplify cleanup of opkg_conf
Move freeing of resources claimed by opkg_config into a dedicated helper
function. This allows reuse in the upcoming two stage argument parsing
and also reduces the numbers of labels needed to perform the actual
freeing of memory.

This change is safe as the individual cleanup methods already check
whether the resource that should be freed has actually been allocated.

Signed-off-by: Philip Lorenz <philip.lorenz@bmw.de>
Signed-off-by: Alex Stewart <alex.stewart@ni.com>